+++ src/resources/web/index.html
@@ -13,7 +13,9 @@
   </main>
 </body>
 <script type="text/javascript" src="mat4.js"></script>
+<script type="text/javascript" src="player.js"></script>
 <script type="text/javascript" src="world.js"></script>
+<script type="text/javascript" src="render.js"></script>
 <script type="text/javascript" src="script.js"></script>
 
 </html>
