+++ src/resources/web/player.ts
@@ -1,0 +1,8 @@
+namespace Player {
+    export let rotY = 0.0;
+    export let rotX = 0.0;
+
+    export let x = 128;
+    export let y = 64;
+    export let z = 128;
+}
